Create STARsolo.R
STARsolo *almost* mimics the 10X format exactly, however it's a blend of CR >= 3 and CR < 3.  There is a features.tsv file, but none of the barcodes.tsv, features.tsv, or matrix.mtx are actually compressed.  This means the default Seurat Read10X function won't work.  This function is just a copy of the Read10X function, with the changes made to support STARsolo format output.